use bare returns
Returns from the lens with the return keyword whenever a problem is found looking up values with the provided keys. Previously this was returning a boolean false, but a boolean false is also a valid value to store in a data structure, so that made it impossible to disambiguate cases where the internal operations of the lens failed versus cases where the lens succeeded but found a boolean false value stored at the target location. Unfortunately, undefined is also a value that can be stored in a data structure! This may demand a stricter solution, but for now we'll consider deliberately storing undefined to be indicative of problems in the code elsewhere.