require immutability function input as lens creation argument
The method interface for injecting a post-processing immutability function was neat, but it requires an extra key on the lens object that is not part of a standard function. This complicates composition, because composed functions will not have that proprietary method, and consequently composed lenses are not quite the same as lenses generated by the factory. Being able to asynchronously inject the immutability is not crucial to the goals of this tool, though -- it is at best inconsequential sparkly polish, and arguably even gratuitious and confusing. Removing this brings us one step closer to lenses as plain functions, which is easier to both understand and compose.