remove getter and setter methods
Removes the hidden methods pointing directly to the getter and setter functions which gets attached to the lens upon creation. The point of including these in the first place was to provide a way to stay true to functional programming pedantry while providing a simpler interface that might be more easy to use. In practice, though, having the simpler interface deviate from bare functions by exposing those methods on the function object was itself a deviation from functional programming pedantry -- it established a proprietary interface that is not immediately recreated by composition. In cases where a getter-only or setter-only function is actually needed, it's still easy to wrap the lens in a function that enforces the desired number of input arguments, especially in ES6 using inline arrow function syntax. Deleting these hidden methods brings us back to a consistent interface -- the lenses are just functions! -- and with it, more predictable behavior with regard to composition.