add composition helper
The lenses are standard functions and compose just fine on their own or with inline user code, but there are some peculiarities which are improved by more careful composition, in particular with regard to standardizing the return values from composed lenses. To help simplify this for users, we'll include an optional composition helper function which handles some of the trickier problems with lens composition.