apply immutability helper in composition function
When a lens is created with a function to enforce immutability added to the hook, it's not enough to simply return the input structure from the composition loop, because we can't be sure that the input structure has been cloned by the immutability function as desired. There is also no longer any reference to the immutability function available, since by that point it only exists inside the closure of the lens that runs it. So the solution is to run each lens on itself, getting the value of the target location in the structure and then immediately setting it. The only purpose of this is to invoke the set function; it will have no effect when data is allowed to be mutable, and will trigger the immutability helper function on set in cases where one has been supplied.