test behavior with complex data types
So far this tool has mostly focused on easily serializable JSON-like structures -- which are probably the most useful, honestly -- but since everything in JavaScript is an object, there's really no reason why we shouldn't use it as an accessor for anything in memory. These additional unit tests make sure it can get and set more complex types such as functions and Date objects, and also set values *on* more complex types, such as adding custom properties. It's probably still a bad idea to extend built in types, but if you can do it with a dot accessor, you should be able to do it with a lens.