allow undefined values
Checking whether values are undefined is not a safe way to determine the execution path, because undefined is a usable value in JavaScript and there may be valid cases where the calling program needs to completely unset an existing value, as opposed to setting it to null. Rather than checking whether an argument is undefined, it's safer to check whether the argument was provided at all by looking at the arguments object.