delete documentation for manual compositions
Manual composition have always been a little unwieldy in their more sophisticated forms, but this crossed a line in the sand once the setter logic was updated to accept undefined as a value to set. After that, it became necessary to check the arguments object in order to distinguish between get and set operations, which means arrow functions can't be used for creating setters, which makes the syntax too long to be appealing. Instead, it's probably better to just provide a quick arrow function example for getters and then point users to the composition function for more sophisticated scenarios.