test for copying complex types should assume early copying
It is possible to use the copy function to do more than duplicate a JSON structure, such as create new Date objects. This test ensures that copying occurs properly with more esoteric JavaScript objects and types. However, it was incorrectly assuming that the input would be mutated, when in fact the goal of the copy function is to leave the input data untouched. This change makes the test look for the changes introduced by the lens on the output data structure instead of on the original input.